Refactored the directory structure of the services, moved backendAPI.dart to the services folder and created smartContract.dart for future use.
