+++ frontend_scv/lib/services/Blockchain/smartContract.dart
@@ -1,0 +1,2 @@
+//This file will contain a class containing methods needed to interact with the smart contract stored on the blockchain.
+//The directory will hold similar classes, e.g. for Metamask communication.
